Avoid garden path
Readers may initially interpret the terms "assistive technology" and
"mouse user" to be the object of these statements. They'll find conflict
with the subsequent term "code path," and have to backtrack to recognize
that the former terms were intended as modifiers.

Rewrite the sentences to reduce the risk of confusion.